requirements.txt
fastapi
uvicorn
pydantic
orjson
python-multipart
python-dotenv
